Optimize player records search and selection
